requirements.txt
python-telegram-bot>=21.0,<22.0
langchain-google-genai>=4.0.0
langgraph>=1.0.0
python-dotenv>=1.0.0
alembic>=1.13.0
sqlalchemy>=2.0.0
